refactor: 인수테스트 합병 (#351)
