+++ backend/src/test/java/com/celuveat/acceptance/auth/MemberAcceptanceTest.java
@@ -1,5 +1,6 @@
 package com.celuveat.acceptance.auth;
 
+import static com.celuveat.acceptance.auth.MemberAcceptanceSteps.로그아웃_요청을_보낸다;
 import static com.celuveat.acceptance.auth.MemberAcceptanceSteps.예상_응답;
 import static com.celuveat.acceptance.auth.MemberAcceptanceSteps.응답을_검증한다;
 import static com.celuveat.acceptance.auth.MemberAcceptanceSteps.회원_탈퇴를_한다;
@@ -29,6 +30,19 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         // then
         응답_상태를_검증한다(응답, 정상_처리);
         응답을_검증한다(응답, 예상_응답);
+    }
+
+    @Test
+    void 로그아웃한다() {
+        // given
+        var 오도 = 멤버("오도");
+        var 세션_아이디 = 회원가입하고_로그인한다(오도);
+
+        // when
+        var 응답 = 로그아웃_요청을_보낸다(세션_아이디, "kakao");
+
+        // then
+        응답_상태를_검증한다(응답, 내용_없음);
     }
 
     @Test
